+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Changelog
 
+## v0.1.2
+
+### Added or Changed
+- `-help` to `--help`
+- `-version` to `--version`
+- `-update` to `--update`
+- Added counters for directories and files
+- Initialized `-s`, `-r`, `-t`, and `-d` options (still not working)
+
 ## v0.1.1
 
 ### Added or Changed
-- Added tree -help
-- Added tree -version
-- Added tree -update
+- Added tree `-help`
+- Added tree `-version`
+- Added tree `-update`
 - Added update.sh
 
 ## v0.1.0